Ignore flake8 warning W503 in our config
According to https://www.flake8rules.com/rules/W503.html the current
rule the code is violating will soon be made the preferred approach.
Rather than changing our code now and change it again later, let's
just ignore this warning. It is just a style issue anyway, so if we
don't catch it in the future it is not a huge issue.

Addresses #15