Convert third party library errors
It is useful to have a error hierarchy that we can assume we can always
catch for our internal library. To make this possible, we need to catch
third party errors and convert them to our own errors. This is also
useful to help understand where the libraries that we call raise errors
to better prepare to handle failure gracefully.

We need to update and add unit tests in order to support these changes
without losing test coverage.

Addresses #12