Switch to yt-dlp
PyTube was not updating their code reliably enough to address YouTube's
frequent changes that would break downloading. We will try yt-dlp as an
alternative and see if it works better.

This requires a non-backwards-compatible API change to the downloaders,
and as a result the CLI and unit tests also need to change. Despite the
non-backwards-compatible changes, keep the major version at 0 to signal
that we are still in development mode.

Addresses #40